+++ crates/nostr/src/prelude.rs
@@ -74,6 +74,7 @@ pub use crate::nips::nip59::{self, *};
 pub use crate::nips::nip62::{self, *};
 pub use crate::nips::nip65::{self, *};
 pub use crate::nips::nip73::{self, *};
+pub use crate::nips::nip88::{self, *};
 pub use crate::nips::nip90::{self, *};
 pub use crate::nips::nip94::{self, *};
 #[cfg(all(feature = "std", feature = "nip96"))]